+++ DarkCodex/Patches/Event_AreaEffects.cs
@@ -14,7 +14,7 @@ using System.Collections.Generic;
 
 namespace DarkCodex
 {
-    [PatchInfo(Severity.Event | Severity.WIP, "Event: Area Effects", "mute player area effects while in dialog", false)]
+    [PatchInfo(Severity.Event, "Event: Area Effects", "mute player area effects while in dialog", false)]
     public class Event_AreaEffects : IDialogStartHandler, IDialogFinishHandler, IPartyCombatHandler, ICutsceneHandler, ICutsceneDialogHandler, IGlobalSubscriber, ISubscriber
     {
         private static readonly List<AreaEffectEntityData> paused = new();